requirements.txt
locustio==0.7.5
pymemcache==1.3.8
